requirements.txt
pymupdf==1.24.10
python-pptx==1.0.2
